requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
filelock==3.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.7
Pillow==9.5.0
platformdirs==2.6.2
psycopg2==2.9.6
psycopg2-binary==2.9.5
Pygments==2.14.0
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1